langchain_modules/image_classifier.py
import logging
import os
import base64
from typing import Dict, Any, Optional
from google.cloud import aiplatform
from google.cloud.aiplatform.gapic.schema import predict

logger = logging.getLogger(__name__)

class RooftopImageClassifier:
    """Class for analyzing rooftop images to estimate solar potential."""
    
    def __init__(self, project_id=None, location=None, endpoint_id=None):
        """Initialize the rooftop image classifier.
        
        Args:
            project_id: Google Cloud project ID
            location: Google Cloud region
            endpoint_id: Vertex AI endpoint ID for the image analysis model
        """
        self.project_id = project_id or os.getenv("VERTEX_PROJECT_ID")
        self.location = location or os.getenv("VERTEX_LOCATION", "us-central1")
        self.endpoint_id = endpoint_id or os.getenv("VERTEX_ENDPOINT_ID")
        
        # Flag to determine if we should use mock responses for demo purposes
        self.mock_mode = os.getenv("MOCK_IMAGE_ANALYSIS", "true").lower() == "true"
    
    def analyze_image(self, image_path: str) -> Dict[str, Any]:
        """Analyze a rooftop image to estimate solar potential.
        
        Args:
            image_path: Path to the image file
            
        Returns:
            Dictionary containing analysis results
        """
        try:
            if self.mock_mode:
                return self._mock_analysis_result(image_path)
            
            # Initialize Vertex AI
            aiplatform.init(project=self.project_id, location=self.location)
            
            # Load and encode image
            with open(image_path, "rb") as f:
                image_content = f.read()
            
            encoded_content = base64.b64encode(image_content).decode("utf-8")
            
            # Create the instance
            instances = [{"image": {"bytesBase64Encoded": encoded_content}}]
            
            # Initialize endpoint
            endpoint = aiplatform.Endpoint(self.endpoint_id)
            
            # Make prediction
            response = endpoint.predict(instances=instances)
            
            # Process response
            return self._process_prediction_response(response)
            
        except Exception as e:
            logger.error(f"Error analyzing image: {str(e)}")
            return {"error": str(e), "suitable": False}
    
    def analyze_image_url(self, image_url: str) -> Dict[str, Any]:
        """Analyze a rooftop image from a URL to estimate solar potential.
        
        Args:
            image_url: URL of the image
            
        Returns:
            Dictionary containing analysis results
        """
        # In a real implementation, this would download the image and then analyze it
        # For now, return mock results
        return self._mock_analysis_result(image_url)
    
    def _process_prediction_response(self, response) -> Dict[str, Any]:
        """Process the prediction response from Vertex AI.
        
        Args:
            response: Response from Vertex AI endpoint
            
        Returns:
            Processed analysis results
        """
        # In a real implementation, this would extract relevant information from the model response
        # For now, returning a mock result
        return {
            "suitable_area_sqm": 25.5,
            "estimated_capacity_kw": 3.8,
            "annual_generation_kwh": 5700,
            "suitable": True,
            "confidence": 0.85,
            "roof_orientation": "south",
            "shading_factor": 0.12
        }
    
    def _mock_analysis_result(self, image_path_or_url: str) -> Dict[str, Any]:
        """Generate a mock analysis result for demonstration purposes.
        
        Args:
            image_path_or_url: Path or URL to the image
            
        Returns:
            Mock analysis results
        """
        # Simple logic to vary the mock results based on the input path hash
        hash_value = sum(ord(c) for c in image_path_or_url) % 100
        
        suitable = hash_value > 20  # 80% chance of being suitable
        
        if suitable:
            area = 20 + (hash_value % 30)  # 20-50 sq meters
            capacity = area * 0.15  # rough estimate of capacity
            generation = capacity * 1500  # rough estimate of annual generation
            shading = (hash_value % 30) / 100  # 0-0.3 shading factor
            confidence = 0.7 + (hash_value % 30) / 100  # 0.7-1.0 confidence
            
            orientations = ["south", "south-west", "south-east", "west", "east"]
            orientation = orientations[hash_value % len(orientations)]
            
            return {
                "suitable_area_sqm": round(area, 1),
                "estimated_capacity_kw": round(capacity, 1),
                "annual_generation_kwh": int(generation),
                "suitable": True,
                "confidence": round(confidence, 2),
                "roof_orientation": orientation,
                "shading_factor": round(shading, 2)
            }
        else:
            reasons = [
                "Excessive shading detected",
                "Roof orientation not optimal",
                "Roof area too small",
                "Complex roof structure"
            ]
            
            return {
                "suitable": False,
                "confidence": round(0.6 + (hash_value % 40) / 100, 2),
                "reason": reasons[hash_value % len(reasons)]
            }

# Create a singleton instance
rooftop_analyzer = RooftopImageClassifier()