requirements.txt
python-telegram-bot==20.6
langchain==0.0.310
google-cloud-aiplatform==1.36.4
python-dotenv==1.0.0
requests==2.31.0
pillow
openai==1.3.7
pydantic==2.4.2
python-uuid==1.30
pandas==2.0.3
matplotlib==3.7.3
numpy==1.24.4
langchain-google-vertexai==0.0.1
langchain-openai==0.0.2
protobuf==4.23.4
google-auth==2.23.3
google-api-core==2.11.1
vertexai==0.0.1
aiohttp==3.8.5
schedule==1.2.1
python-telegram-bot[webhooks]==20.6